Moved GtQuerymatchoptions to own file.
