+++ src/match/querymatch-align.h
@@ -1,0 +1,59 @@
+/*
+  Copyright (c) 2015 Stefan Kurtz <kurtz@zbh.uni-hamburg.de>
+  Copyright (c) 2015 Center for Bioinformatics, University of Hamburg
+
+  Permission to use, copy, modify, and distribute this software for any
+  purpose with or without fee is hereby granted, provided that the above
+  copyright notice and this permission notice appear in all copies.
+
+  THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+  WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
+  MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+  ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+  WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
+  ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+  OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+*/
+
+#ifndef QUERYMATCH_ALIGN_H
+#define QUERYMATCH_ALIGN_H
+
+#include "core/types_api.h"
+#include "match/ft-front-prune.h"
+
+typedef struct GtQuerymatchoutoptions GtQuerymatchoutoptions;
+
+GtQuerymatchoutoptions *gt_querymatchoutoptions_new(
+                                GtUword alignmentwidth,
+                                GtUword errorpercentage,
+                                GtUword maxalignedlendifference,
+                                GtUword history,
+                                GtUword perc_mat_history,
+                                GtExtendCharAccess extend_char_access,
+                                GtUword sensitivity);
+
+void gt_querymatchoutoptions_delete(
+        GtQuerymatchoutoptions *querymatchoutoptions);
+
+void gt_querymatchoutoptions_set_seed(
+                            GtQuerymatchoutoptions *querymatchoutoptions,
+                            GtUword pos1,GtUword pos2,GtUword len);
+
+void gt_querymatch_alignment_prepare(GtQuerymatchoutoptions
+                                     *querymatchoutoptions,
+                                     const GtEncseq *encseq,
+                                     GtUword dbstart,
+                                     GtUword dblen,
+                                     GtUword querystartabsolute,
+                                     GtUword querylen,
+                                     GtUword edist,
+                                     GT_UNUSED bool greedyextension);
+
+void gt_querymatchoutoptions_alignment_show(const GtQuerymatchoutoptions
+                                              *querymatchoutoptions);
+
+void gt_querymatchoutoptions_exact_alignment_show(const GtQuerymatchoutoptions
+                                                     *querymatchoutoptions,
+                                                  GtUword len);
+
+#endif
